refactor: move templating functions into it's own package
